docker/python-oscillogram-analyzer/pyproject.toml
[tool.poetry]
name = "oscillogram-analyzer"
version = "0.1.0"
description = "решение кейса ПАО 'Россети'"
authors = ["Flow Team <flow@team.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
injector = "^0.18.4"
click = "^7.1.2"
kafka-python = "^2.0.2"
comtrade = "^0.0.1"
matplotlib = "^3.3.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"

[build-system]
requires = ["poetry>=1.1.2"]
build-backend = "poetry.masonry.api"
